+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/account/SocialLoginAPI.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.infrastructure.web.account;
 
-import com.mobble.mobbleserver.account.auth.dto.request.SocialLoginRequestDto;
-import com.mobble.mobbleserver.account.auth.dto.response.SocialLoginResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.account.dto.request.SocialLoginRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.account.dto.response.SocialLoginResponseDto;
 import com.mobble.mobbleserver.account.auth.service.SocialLoginService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
