+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/account/dto/request/SignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.auth.dto.request;
+package com.mobble.mobbleserver.infrastructure.web.account.dto.request;
 
 import com.mobble.mobbleserver.account.auth.oauth.verifier.dto.SocialUserInfo;
 import com.mobble.mobbleserver.domain.ground.Ground;
